+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
   - Validate AccessPass before client connection ([#1356](https://github.com/malbeclabs/doublezero/issues/1356))
 - Client
   - Add initial route liveness probing, initially disabled for rollout
+  - Add route liveness prometheus metrics
 
 ## [v0.6.10](https://github.com/malbeclabs/doublezero/compare/client/v0.6.9...client/v0.6.10) – 2025-11-05